# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Tahir AI Writer API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /users/{id}/reset-password:
    post:
      operationId: resetUserPassword
      tags: [users]
      summary: Reset any user's password (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /sites:
    get:
      operationId: listSites
      tags: [sites]
      summary: List sites for current user
      parameters:
        - name: folderId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of sites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Site"
    post:
      operationId: createSite
      tags: [sites]
      summary: Add a new site
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSiteRequest"
      responses:
        "201":
          description: Site created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
  /sites/bulk-add:
    post:
      operationId: bulkAddSites
      tags: [sites]
      summary: Bulk add multiple sites
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkAddSitesRequest"
      responses:
        "200":
          description: Sites added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkAddSitesResponse"
  /sites/auto-connect:
    post:
      operationId: autoConnectSite
      tags: [sites]
      summary: Auto-connect a WordPress site by URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutoConnectRequest"
      responses:
        "200":
          description: Site connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
        "400":
          description: Connection failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /sites/{id}:
    get:
      operationId: getSite
      tags: [sites]
      summary: Get site by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Site
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateSite
      tags: [sites]
      summary: Update site
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSiteRequest"
      responses:
        "200":
          description: Site updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
    delete:
      operationId: deleteSite
      tags: [sites]
      summary: Delete site
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /sites/{id}/test:
    post:
      operationId: testSite
      tags: [sites]
      summary: Test site connection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteTestResult"
  /sites/{id}/pin:
    patch:
      operationId: toggleSitePin
      tags: [sites]
      summary: Toggle pin on a site
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Site updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, author]
        permissions:
          $ref: "#/components/schemas/UserPermissions"
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - role
        - createdAt
    UserPermissions:
      type: object
      properties:
        canGenerate:
          type: boolean
        canPublish:
          type: boolean
        canSchedule:
          type: boolean
        canManageSites:
          type: boolean
        maxSites:
          type: ["integer", "null"]
        maxArticlesPerDay:
          type: ["integer", "null"]
      required:
        - canGenerate
        - canPublish
        - canSchedule
        - canManageSites
    CreateUserRequest:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, author]
        permissions:
          $ref: "#/components/schemas/UserPermissions"
      required:
        - email
        - name
        - password
        - role
    UpdateUserRequest:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, author]
        permissions:
          $ref: "#/components/schemas/UserPermissions"
...
    ResetPasswordRequest:
      type: object
      properties:
        newPassword:
          type: string
      required:
        - newPassword
    Site:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        folderId:
          type: ["integer", "null"]
        name:
          type: string
        url:
          type: string
        username:
          type: string
        status:
          type: string
          enum: [connected, disconnected, pending]
        isPinned:
          type: boolean
        category:
          type: string
        lastTested:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - userId
        - name
        - url
        - username
        - status
        - isPinned
        - createdAt
    CreateSiteRequest:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        username:
          type: string
        applicationPassword:
          type: string
        folderId:
          type: ["integer", "null"]
        category:
          type: string
      required:
        - url
        - username
        - applicationPassword
    UpdateSiteRequest:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        username:
          type: string
        applicationPassword:
          type: string
        folderId:
          type: ["integer", "null"]
        category:
          type: string
        isPinned:
          type: boolean
    BulkAddSitesRequest:
      type: object
      properties:
        sites:
          type: array
          items:
            $ref: "#/components/schemas/CreateSiteRequest"
      required:
        - sites
    BulkAddSitesResponse:
      type: object
      properties:
        added:
          type: integer
        failed:
          type: integer
        sites:
          type: array
          items:
            $ref: "#/components/schemas/Site"
      required:
        - added
        - failed
        - sites
    AutoConnectRequest:
      type: object
      properties:
        url:
          type: string
        folderId:
          type: ["integer", "null"]
      required:
        - url
    SiteTestResult:
      type: object
      properties:
        siteId:
          type: integer
        success:
          type: boolean
        message:
          type: string
      required:
        - siteId
        - success
        - message
...