# lib/api-zod/src/generated/types/bulkAddSitesRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tahir AI Writer API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateSiteRequest } from "./createSiteRequest";

export interface BulkAddSitesRequest {
  sites: CreateSiteRequest[];
}